requirements.txt
absl-py==1.4.0
altair==5.5.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
backcall==0.2.0
BareNecessities==0.2.8
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.7.0
CacheControl==0.12.5
cachetools==5.3.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cycler==0.10.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.110.0
fastjsonschema==2.20.0
filelock==3.13.1
firebase-admin==3.2.0
Flask==2.2.2
Flask-Cors==3.0.8
Flask-Mail==0.9.1
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==1.34.1
google-api-python-client==2.122.0
google-auth==2.16.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.8.4
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.51.1
grpcio-status==1.48.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.35.1
importlib-metadata==6.0.0
ipykernel==6.29.5
ipython==8.30.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.4
joblib==1.2.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.5.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==15.0.6.1
librosa==0.10.1
llvmlite==0.44.0
Mail==2.1.0
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.0.8
mysql-connector-python==9.0.0
namex==0.0.7
narwhals==1.24.1
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
notebook==7.3.1
notebook_shim==0.2.4
npm==0.1.1
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.0
opt-einsum==3.3.0
optional-django==0.1.0
optree==0.10.0
overrides==7.7.0
packaging==24.1
pandas==2.1.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.3.6
pooch==1.8.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.23.0
protobuf==3.20.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.1
Pygments==2.18.0
pyngrok==7.2.8
pyparsing==3.0.9
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2022.7.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
QtPy==2.3.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.0
seaborn==0.13.0
Send2Trash==1.8.3
shapely==2.0.6
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.36.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.2.0
terminado==0.18.1
threadpoolctl==3.1.0
tifffile==2024.8.28
tinycss2==1.3.0
toml==0.10.2
torch==2.0.0
torchvision==0.15.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.29.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
widgetsnbextension==4.0.13
wrapt==1.14.1
xlrd==2.0.1
zipp==3.12.1
